+++ exporters/stdout/stdoutlog/record.go
@@ -4,8 +4,6 @@
 package stdoutlog // import "go.opentelemetry.io/otel/exporters/stdout/stdoutlog"
 
 import (
-	"encoding/json"
-	"errors"
 	"time"
 
 	"go.opentelemetry.io/otel/attribute"
@@ -16,105 +14,6 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
-func newValue(v attribute.Value) value {
-	return value{Value: v}
-}
-
-type value struct {
-	attribute.Value
-}
-
-// MarshalJSON implements a custom marshal function to encode attribute.Value.
-func (v value) MarshalJSON() ([]byte, error) {
-	var jsonVal struct {
-		Type  string
-		Value any
-	}
-	jsonVal.Type = valueType(v.Value)
-
-	switch v.Type() {
-	case attribute.STRING:
-		jsonVal.Value = v.AsString()
-	case attribute.INT64:
-		jsonVal.Value = v.AsInt64()
-	case attribute.FLOAT64:
-		jsonVal.Value = v.AsFloat64()
-	case attribute.BOOL:
-		jsonVal.Value = v.AsBool()
-	case attribute.BYTESLICE:
-		jsonVal.Value = v.AsByteSlice()
-	case attribute.BOOLSLICE:
-		jsonVal.Value = v.AsBoolSlice()
-	case attribute.INT64SLICE:
-		jsonVal.Value = v.AsInt64Slice()
-	case attribute.FLOAT64SLICE:
-		jsonVal.Value = v.AsFloat64Slice()
-	case attribute.STRINGSLICE:
-		jsonVal.Value = v.AsStringSlice()
-	case attribute.MAP:
-		m := v.AsMap()
-		values := make([]keyValue, 0, len(m))
-		for _, kv := range m {
-			values = append(values, keyValue{
-				Key:   string(kv.Key),
-				Value: newValue(kv.Value),
-			})
-		}
-
-		jsonVal.Value = values
-	case attribute.SLICE:
-		s := v.AsSlice()
-		values := make([]value, 0, len(s))
-		for _, e := range s {
-			values = append(values, newValue(e))
-		}
-
-		jsonVal.Value = values
-	case attribute.EMPTY:
-		jsonVal.Value = nil
-	default:
-		return nil, errors.New("invalid attribute.Type")
-	}
-
-	return json.Marshal(jsonVal)
-}
-
-func valueType(v attribute.Value) string {
-	switch v.Type() {
-	case attribute.EMPTY:
-		return "Empty"
-	case attribute.BOOL:
-		return "Bool"
-	case attribute.INT64:
-		return "Int64"
-	case attribute.FLOAT64:
-		return "Float64"
-	case attribute.STRING:
-		return "String"
-	case attribute.BYTESLICE:
-		return "Bytes"
-	case attribute.SLICE:
-		return "Slice"
-	case attribute.MAP:
-		return "Map"
-	case attribute.BOOLSLICE:
-		return "BoolSlice"
-	case attribute.INT64SLICE:
-		return "Int64Slice"
-	case attribute.FLOAT64SLICE:
-		return "Float64Slice"
-	case attribute.STRINGSLICE:
-		return "StringSlice"
-	default:
-		return "Invalid"
-	}
-}
-
-type keyValue struct {
-	Key   string
-	Value value
-}
-
 // recordJSON is a JSON-serializable representation of a Record.
 type recordJSON struct {
 	Timestamp         *time.Time `json:",omitempty"`
@@ -122,8 +21,8 @@ type recordJSON struct {
 	EventName         string     `json:",omitempty"`
 	Severity          log.Severity
 	SeverityText      string
-	Body              value
-	Attributes        []keyValue
+	Body              attribute.Value
+	Attributes        []attribute.KeyValue
 	TraceID           trace.TraceID
 	SpanID            trace.SpanID
 	TraceFlags        trace.TraceFlags
@@ -138,13 +37,13 @@ func (e *Exporter) newRecordJSON(r sdklog.Record) recordJSON {
 		EventName:    r.EventName(),
 		Severity:     r.Severity(),
 		SeverityText: r.SeverityText(),
-		Body:         newValue(r.Body()),
+		Body:         r.Body(),
 
 		TraceID:    r.TraceID(),
 		SpanID:     r.SpanID(),
 		TraceFlags: r.TraceFlags(),
 
-		Attributes: make([]keyValue, 0, r.AttributesLen()),
+		Attributes: make([]attribute.KeyValue, 0, r.AttributesLen()),
 
 		Resource: res,
 		Scope:    r.InstrumentationScope(),
@@ -153,10 +52,7 @@ func (e *Exporter) newRecordJSON(r sdklog.Record) recordJSON {
 	}
 
 	r.WalkAttributes(func(kv attribute.KeyValue) bool {
-		newRecord.Attributes = append(newRecord.Attributes, keyValue{
-			Key:   string(kv.Key),
-			Value: newValue(kv.Value),
-		})
+		newRecord.Attributes = append(newRecord.Attributes, kv)
 		return true
 	})